1. Renamed the file for the SettingsPopUp to follow convention 2. Brought a lot of context functionalities and elements to this level 3. Updated the data state handling and composition in order to effectively change room settings with visual conditional component rendering 4. Added useEffect cleanup handling in the case there IS an actual change to the state of the data relative to the previous value 5. Created a handleSave function that handles the saving of the new state of the room settings
- Makes a post request to the /api/rooms/update/:roomUID route with the new difficulty, topics, and lobbyAccess
- This request returns an updated variable, which if set to true:
	- Effects the state to reflects a successful POST request
	- Tracking of the changes of the current room settings and the temporary ones in the settings pop up
	- Communicates any possible changes to the backend socket.io event listener 'backendSettingsChange' with room uid, user name, and socket id
	- Closes the pop up
	- Makes a toast for the client sender!

6. Added defaultValue, value, and onChange handling for the multi-select Select tag for the topics
7. Added the defaultValue, value, and onChange handling for the multi-select Select tag for the difficulty

Learned: Had to bring the socket event listener 'frontendSettingsChange' up a level as I was trying to do a socket.on listen when I was going to cleanup via useEffect on component dismount or when I closed the popup via the boolean state value passed into the component, but this effectively does the same and there would be no awaiting this call, therefore it was unreliable to leave in this component! Had to bring it up as the header is never un-rendered unlike the SettingsPopUp. The more you know! (small details)